issue-#2
add lastAccessed time to replace modified time
add setting of lastAccessed time when loading the session so that when
in use it does not expire; previously it was only updated upon save.
update to use time.Now().UTC() just in case this code is running on
multiple machines (load balanced application) that may not have the
same timezone this will still function based on the UTC timezone.